+++ src/createComputerNumbers.js
@@ -1,4 +1,4 @@
-function makeComputerNum(){
+export function createComputerNumbers(){
     let tmp_num = MissionUtils.Random.pickNumberInRange(1, 9);
     let set = new Set([tmp_num]);
 
@@ -9,6 +9,3 @@ function makeComputerNum(){
     const computer_nums = [... set]
     return computer_nums.join('');
 }
-
-
-export default makeComputerNum;
